// src/interfaces/jmt.interface.ts
export interface coverData {
	engineLogoOn: boolean
	info: coverInfo
}

export interface sceneObj {
	tours: any[]
	views: any[]
}
export interface coverInfo {
	sceneObjs: sceneObj
	isLikes: boolean
	isMessage: boolean
	isShare: boolean
	danmu: boolean
	browseService: boolean
	buttonOptions: coverButtonOptions
	closeMusic: null | boolean
	contact: coverContact
	custService: boolean
	custServiceCode: null | string
	customLogo: null | string
	descFlag: boolean
	description: string
	durationEndTs: null | string
	enLang: boolean
	extobjs: any[]
	floor: number
	hideLogo: boolean
	password: string
	liveService: boolean
	loadingThumb: null | string
	loadingVideo: null | string
	lupt: number
	lut: null | string
	lutName: null | string
	lutid: null | string
	lutsz: null | string
	minimap: coverMinimap
	mobileThumb: null | string
	modelRotate: boolean
	musicAutoPlay: boolean
	musicFile: null | string
	musicId: null | string
	musicName: null | string
	myBrowseService: boolean
	myCustService: boolean
	name: string
	panoImg: null | string
	pluginFlag: boolean
	renderOver: boolean
	sceneId: string
	sceneShare: null | string
	sceneTemplateLive: coverSceneTemplateLive
	scripts: string[]
	showDescFlag: number
	startCamera: string
	tempid: string
	thumb: string
	titleFlag: boolean
	token: null | string
	typeId: string
	usePwd: boolean
	useThumbLoading: boolean
	username: string
	viewCount: number
	openingVideo: {
		show: boolean
		url: string
		name: string
		showSkip: boolean
	}
	openTour: {
		show: boolean
		tour: string
		changeBGM: boolean
	}
	outline: {
		show: boolean
	}
	guidePage: {
		show: boolean
	}
	customButton: {
		visible: boolean
	}
	spotSwitchBtn: {
		show: boolean
	}
}

export interface coverButtonOptions {
	comment: boolean
	headerLink: boolean
	hideAll: boolean
	hideHeader: boolean
	share: boolean
}

export interface coverContact {
	contactAddress: string
	contactEmail: null | string
	contactName: string
	contactPhone: string
	externalUrl: null | string
	headimgurl: string
	userId: string
	wxQrcode: null | string
}

export interface coverMinimap {
	mapImage: string
	rect: number
	show: boolean
	position: {
		x: number
		y: number
	}
}

export interface coverRect {
	x0: number
	x1: number
	y0: number
	y1: number
}

export interface coverSceneTemplateLive {
	liveEndTime: null | string
	liveOpenType: number
	liveStartTime: null | string
	liveState: boolean
	liveUrl: null | string
	senceTemplateId: null | string
}

export interface devicesData {
	id: string
	type: string
	description: string
}

export interface renderParams {
	cameraPosition: number[]
	cameraRotation: number[]
	cameraFov: number
}
// ...
export interface bakeData {
	floodDarkLimit: number
	bgMapGamma: number
	device: string
	useBgAo: boolean
	width: number
	maxDirectSample: number
	maxIndirectSample: number
	bgStrength: number
	bgColor: number[]
	bgAoDistance: number
	usePostProcessFilters: boolean
	sampleCount: number
	bounceCount: number
	bgAoFactor: number
	lightmapResolution: number
	height: number
	useBg: boolean
	useBgMap: boolean
	devices: string[]
	tileSize: number
	exportLightmapUvs: boolean
}

export interface jobsData {
	stateTag: string
	jobs: JobItem[]
}

export interface JobItem {
	id: number
	type: string
	sceneName: string
	state: string
	timestamp: number
	runningTimestamp: number
	executedTimestamp: number | null
	now: number
	message: null
	progress: JobProgress
	info: any[]
}

export interface JobProgress {
	message: string
	total: number | null
	done: number | null
	elapsed: number | null
}

export interface sceneData {
	id: number
}